allow for dragging keywords
